Moved memory type to shared module
