+++ android/app/src/main/kotlin/pl/kwasow/managers/MemoriesManagerImpl.kt
@@ -1,6 +1,6 @@
 package pl.kwasow.managers
 
-import pl.kwasow.data.types.Memory
+import pl.kwasow.flamingo.types.Memory
 import java.time.LocalDate
 
 class MemoriesManagerImpl(
