+++ android/app/src/main/kotlin/pl/kwasow/managers/SystemManager.kt
@@ -1,7 +1,7 @@
 package pl.kwasow.managers
 
-import pl.kwasow.data.types.Memory
 import pl.kwasow.data.types.User
+import pl.kwasow.flamingo.types.Memory
 
 interface SystemManager {
     // ====== Methods
