+++ android/app/src/main/kotlin/pl/kwasow/ui/widgets/memories/MemoriesWidgetViewModel.kt
@@ -6,7 +6,7 @@ import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.launch
-import pl.kwasow.data.types.Memory
+import pl.kwasow.flamingo.types.Memory
 import pl.kwasow.managers.MemoriesManager
 
 class MemoriesWidgetViewModel(
