+++ android/build.gradle.kts
@@ -7,4 +7,5 @@ plugins {
     alias(libs.plugins.firebase.crashlytics) apply false
     alias(libs.plugins.google.mapsplatform) apply false
     alias(libs.plugins.google.services) apply false
+    alias(libs.plugins.jetbrains.kotlin.jvm) apply false
 }
