+++ android/gradle/libs.versions.toml
@@ -39,6 +39,7 @@ googleProtobufPlugin = "0.9.4"
 googleServicesPlugin = "4.4.2"
 kotlinPlugins = "2.1.10"
 ktlintPlugin = "12.2.0"
+jetbrainsKotlinJvm = "2.1.10"
 
 [libraries]
 android-credentials-base = { group = "androidx.credentials", name = "credentials", version.ref = "androidCredentials" }
@@ -95,3 +96,4 @@ kotlin-android = { id = "org.jetbrains.kotlin.android", version.ref = "kotlinPlu
 kotlin-compose = { id = "org.jetbrains.kotlin.plugin.compose", version.ref = "kotlinPlugins" }
 kotlin-serialization = { id = "org.jetbrains.kotlin.plugin.serialization", version.ref = "kotlinPlugins" }
 ktlint = { id = "org.jlleitschuh.gradle.ktlint", version.ref = "ktlintPlugin" }
+jetbrains-kotlin-jvm = { id = "org.jetbrains.kotlin.jvm", version.ref = "jetbrainsKotlinJvm" }
